+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 1.0.12 - DATE
+## 1.0.13 - DATE
+
+
+
+## 1.0.12 - 26/10/2023
+
+### Fixed
+
+- Adapted the plugin following the assessment by the WordPress Plugin Review team.
 
 ### Improvements